+++ NEWS.md
@@ -1,5 +1,7 @@
 # calcite 0.1.2
 
+- `calcite_button()` click handler can be tracked via `input$your_button$clicked`
+
 # calcite 0.1.1
 
 - Add event handler for clicks in `calcite_button()`. See example `system.file("examples/button-click-test.R", package = "calcite")`